requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
daphne==4.2.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
dj-database-url==3.0.1
whitenoise==6.9.0
django-redis==5.4.0
social-auth-app-django==5.4.1
djangorestframework-simplejwt==5.5.1
django-filter==24.3
drf-yasg==1.21.10
channels-redis==4.3.0
google-auth==2.40.3
twilio==9.3.3
agora-token-builder==1.0.0
sentry-sdk==2.37.0
django-prometheus==2.4.1
gunicorn==23.0.0
uvicorn==0.30.0

django-allauth
dj-rest-auth[with_social]

